pyproject.toml
[project]
name = "agentic-token-bench"
version = "0.1.0"
description = "Benchmark and plugin framework for measuring token-saving tools in agentic coding"
readme = "README.md"
requires-python = ">=3.12"
license = "Apache-2.0"
authors = [{name = "Patrick McFadin"}]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "typer>=0.15",
    "pydantic>=2.0",
    "duckdb>=1.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.urls]
Repository = "https://github.com/pmcfadin/agentic-token-bench"
Issues = "https://github.com/pmcfadin/agentic-token-bench/issues"

[project.scripts]
atb = "benchmarks.harness.cli:app"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-xdist>=3.0",
    "ruff>=0.9",
    "jsonschema>=4.0",
]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-n 1"
markers = [
    "smoke: smoke tests for CI",
    "integration: tests requiring real agent CLI binaries",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["benchmarks", "agents", "tools"]
